Remove unused code previously used to generate JSON schema
