--- POPULAR.json
@@ -1,65 +1,0 @@
-{
-  "Google": [
-    "Pixel 3",
-    "Pixel 3a",
-    "Pixel 3a XL",
-    "Pixel 3 XL",
-    "Pixel 4",
-    "Pixel 4 XL",
-    "Pixel 5"
-  ],
-  "Huawei": [
-    "Nexus 6P",
-    "Honor View 10",
-    "Honor 7X"
-  ],
-  "LGE": [
-    "LG G5",
-    "LG G6",
-    "Nexus 4",
-    "Nexus 5",
-    "Nexus 5X",
-    "LG V50 ThinQ"
-  ],
-  "Motorola": [
-    "Nexus 6",
-    "Moto Z"
-  ],
-  "OnePlus": [
-    "OnePlus 7",
-    "OnePlus 7 Pro",
-    "OnePlus 8",
-    "OnePlus 9 5G"
-  ],
-  "Samsung": [
-    "Galaxy S7",
-    "Galaxy S7 Edge",
-    "Galaxy S8",
-    "Galaxy S8+",
-    "Galaxy S9",
-    "Galaxy S9+",
-    "Galaxy S10",
-    "Galaxy S10e",
-    "Galaxy S10+",
-    "Galaxy S20",
-    "Galaxy S20+",
-    "Galaxy S20 5G",
-    "Galaxy S20 FE 5G",
-    "Galaxy S21+ 5G",
-    "Galaxy S21 Ultra 5G",
-    "Galaxy Note20",
-    "Galaxy Note20 5G"
-  ],
-  "Nokia": [
-    "Nokia 3.1 Plus",
-    "Nokia 7.1"
-  ],
-  "Sony": [
-    "Xperia 1",
-    "Xperia 10",
-    "Xperia 10 Plus",
-    "Xperia XZ3",
-    "Xperia XZ2",
-    "Xperia XZ2 Premium"
-  ]
-}
